fix: change basic server files structure.
